Move nodes into separate files
